+++ packages/alfa-media/src/feature/feature.ts
@@ -4,6 +4,7 @@ import {
   Number,
   Percentage,
   String,
+  Token,
 } from "@siteimprove/alfa-css";
 import { Device } from "@siteimprove/alfa-device";
 import { Hash } from "@siteimprove/alfa-hash";
@@ -13,8 +14,13 @@ import * as json from "@siteimprove/alfa-json";
 
 import { Media } from "../media";
 import { Resolver } from "../resolver";
+import { Err, Ok, Result } from "@siteimprove/alfa-result";
+import { Parser } from "@siteimprove/alfa-parser";
+import { Slice } from "@siteimprove/alfa-slice";
 
 const { isLength } = Length;
+
+const { delimited, either, map, mapResult, option, pair, right } = Parser;
 
 /**
  * @see https://drafts.csswg.org/mediaqueries/#media-feature
@@ -123,6 +129,8 @@ export class Feature<
   }
 }
 
+import { Orientation } from "./orientation";
+
 export namespace Feature {
   export interface JSON {
     [key: string]: json.JSON;
@@ -145,4 +153,90 @@ export namespace Feature {
   export function isFeature(value: unknown): value is Feature {
     return value instanceof Feature;
   }
+  export const { fromValue: orientation, isOrientation } = Orientation;
+
+  export function from(
+    name: string,
+    value: Option<Feature.Value>
+  ): Result<Feature, string> {
+    switch (name) {
+      case "orientation":
+        return orientation(value);
+      default:
+        return Err.of("Unknown feature");
+    }
+  }
+
+  /**
+   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-name
+   */
+  const parseName = map(Token.parseIdent(), (ident) =>
+    ident.value.toLowerCase()
+  );
+
+  /**
+   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-value
+   */
+  const parseValue = either(
+    either(
+      map(Token.parseNumber(), (number) => Number.of(number.value)),
+      map(Token.parseIdent(), (ident) => String.of(ident.value.toLowerCase()))
+    ),
+    either(
+      map(
+        pair(
+          Token.parseNumber((number) => number.isInteger),
+          right(
+            delimited(option(Token.parseWhitespace), Token.parseDelim("/")),
+            Token.parseNumber((number) => number.isInteger)
+          )
+        ),
+        (result) => {
+          const [left, right] = result;
+
+          return Percentage.of(left.value / right.value);
+        }
+      ),
+      Length.parse
+    )
+  );
+
+  /**
+   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-plain
+   */
+  const parsePlain: Parser<Slice<Token>, Feature, string> = mapResult(
+    pair(
+      parseName,
+      right(
+        delimited(option(Token.parseWhitespace), Token.parseColon),
+        parseValue
+      )
+    ),
+    (result) => {
+      const [name, value] = result;
+
+      return name === "orientation"
+        ? from(name, Option.of(value))
+        : (Ok.of(Feature.of(name, Option.of(value))) as Result<
+            Feature,
+            string
+          >);
+    }
+  );
+
+  /**
+   * @see https://drafts.csswg.org/mediaqueries/#typedef-mf-boolean
+   */
+  const parseBoolean = map(parseName, (name) =>
+    Feature.of<string, never>(name)
+  );
+
+  /**
+   * @see https://drafts.csswg.org/mediaqueries/#typedef-media-feature
+   */
+  export const parse = delimited(
+    Token.parseOpenParenthesis,
+    delimited(option(Token.parseWhitespace), either(parsePlain, parseBoolean)),
+    Token.parseCloseParenthesis
+  );
 }
