+++ packages/alfa-media/tsconfig.json
@@ -7,9 +7,8 @@
     "src/condition/index.ts",
     "src/condition/negation.ts",
     "src/feature/feature.ts",
-    "src/feature/foo.ts",
-    "src/feature/orientation.ts",
     "src/feature/index.ts",
+    "src/feature/orientation.ts",
     "src/index.ts",
     "src/media.ts",
     "src/query.ts",
